requirements.txt
streamlit
streamlit-authenticator
st-theme
pandas
numpy
altair
emoji
google-cloud-storage
